mahak-gupta/requirements.txt
Pillow==9.1.1
requests==2.27.1
streamlit==1.9.0
streamlit_lottie==0.0.3
